pyproject.toml
[build-system]
requires = [
  "setuptools>=68",
  "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "milksnake"
version = "0.1.0"
description = "SNMP simulator that responds based on provided MIB files"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Stanislaw Borowy", email = "stabor@startmail.com" },
    { name = "Przemyslaw Maresz" }
]
license = { text = "GPL-3.0-or-later" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Operating System :: OS Independent",
]

scripts = { "milksnake" = "milksnake.main:main" }

# If you prefer to keep runtime dependencies in requirements.txt, you can omit this.
# Listing them here makes builds self-contained per PEP 621.
dependencies = [
  "pyasn1==0.6.1",
  "pysnmp==7.1.22",
  "pyyaml>=6.0.3",
  "sortedcontainers>=2.4.0",
]


[tool.setuptools]
# Use package discovery for the src layout or plain package folder.
[tool.setuptools.packages.find]
where = ["."]
include = ["milksnake*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
    "pytest-asyncio>=1.3.0",
    "black>=25.11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --cov=milksnake --cov-report=term-missing --cov-report=xml"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",    # pycodestyle errors
  "W",    # pycodestyle warnings  
  "F",    # pyflakes
  "I",    # isort
  "N",    # pep8-naming
  "UP",   # pyupgrade
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "SIM",  # flake8-simplify
  "ANN",  # flake8-annotations
]

extend-ignore = [
  "ANN401", # allow Any type annotations
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN"]
